Let's move integration tests that are using mocks to unit tests
