+++ camel-sap/camel-sap-component/data/testDocument.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="ASCII"?>
-<idoc:Document xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:TEST_IDOC_TYPE-TEST_IDOC_TYPE_EXTENSION-TEST_SYSTEM_RELEASE-TEST_APPLICATION_RELEASE="http://sap.fusesource.org/idoc/TEST_REPOSITORY/TEST_IDOC_TYPE/TEST_IDOC_TYPE_EXTENSION/TEST_SYSTEM_RELEASE/TEST_APPLICATION_RELEASE" xmlns:idoc="http://sap.fusesource.org/idoc" archiveKey="archiveKeyValue" client="clientValue" creationDate="1861-04-12T11:50:37.722-0500" creationTime="2014-07-29T04:30:15.725-0400" direction="directionValue" EDIMessage="ediMessageValue" EDIMessageGroup="editMessageGroupValue" EDIMessageType="editMessageTypeValue" EDIStandardFlag="ediStandardFlagValue" EDIStandardVersion="ediStandardVersionValue" EDITransmissionFile="ediTransmissionFileValue" iDocCompoundType="idocCompoundTypeValue" iDocNumber="idocNumberValue" iDocSAPRelease="idocSAPReleaseValue" iDocType="idocTypeValue" iDocTypeExtension="idocTypeExtensionValue" messageCode="messageCodeValue" messageFunction="messageFunctionValue" messageType="messageTypeValue" outputMode="outputModeValue" recipientAddress="recipientAddressValue" recipientLogicalAddress="recipientLogicalAddressValue" recipientPartnerFunction="recipientPartnerFunctionValue" recipientPartnerNumber="recipientPartnerNumberValue" recipientPartnerType="recipientPartnerTypeValue" recipientPort="recipientPortValue" senderAddress="senderAddressValue" senderLogicalAddress="senderLogicalAddressValue" senderPartnerFunction="senderPartnerFunctionValue" senderPartnerNumber="senderPartnerNumberValue" senderPartnerType="senderPartnerTypeValue" senderPort="senderPortValue" serialization="serializationValue" status="statusValue" testFlag="testFlagValue">
-  <rootSegment xsi:type="TEST_IDOC_TYPE-TEST_IDOC_TYPE_EXTENSION-TEST_SYSTEM_RELEASE-TEST_APPLICATION_RELEASE:ROOT" document="/">
+<idoc:Document xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:TEST_IDOC_TYPE-TEST_IDOC_TYPE_EXTENSION-TEST_SYSTEM_VERSION-TEST_APPLICATION_VERSION="http://sap.fusesource.org/idoc/TEST_REPOSITORY/TEST_IDOC_TYPE/TEST_IDOC_TYPE_EXTENSION/TEST_SYSTEM_VERSION/TEST_APPLICATION_VERSION" xmlns:idoc="http://sap.fusesource.org/idoc" archiveKey="archiveKeyValue" client="clientValue" creationDate="1861-04-12T12:15:11.632+0100" creationTime="2023-03-28T16:30:15.632+0200" direction="directionValue" EDIMessage="ediMessageValue" EDIMessageGroup="editMessageGroupValue" EDIMessageType="editMessageTypeValue" EDIStandardFlag="ediStandardFlagValue" EDIStandardVersion="ediStandardVersionValue" EDITransmissionFile="ediTransmissionFileValue" iDocCompoundType="idocCompoundTypeValue" iDocNumber="idocNumberValue" iDocSAPRelease="idocSAPReleaseValue" iDocType="idocTypeValue" iDocTypeExtension="idocTypeExtensionValue" messageCode="messageCodeValue" messageFunction="messageFunctionValue" messageType="messageTypeValue" outputMode="outputModeValue" recipientAddress="recipientAddressValue" recipientLogicalAddress="recipientLogicalAddressValue" recipientPartnerFunction="recipientPartnerFunctionValue" recipientPartnerNumber="recipientPartnerNumberValue" recipientPartnerType="recipientPartnerTypeValue" recipientPort="recipientPortValue" senderAddress="senderAddressValue" senderLogicalAddress="senderLogicalAddressValue" senderPartnerFunction="senderPartnerFunctionValue" senderPartnerNumber="senderPartnerNumberValue" senderPartnerType="senderPartnerTypeValue" senderPort="senderPortValue" serialization="serializationValue" status="statusValue" testFlag="testFlagValue">
+  <rootSegment xsi:type="TEST_IDOC_TYPE-TEST_IDOC_TYPE_EXTENSION-TEST_SYSTEM_VERSION-TEST_APPLICATION_VERSION:ROOT" document="/">
     <segmentChildren parent="//@rootSegment">
-      <LEVEL1 parent="//@rootSegment" document="/" FIELD0="FIELD0_VALUE" FIELD1="FIELD1_VALUE" FIELD2="FIELD2_VALUE" FIELD3="FIELD3_VALUE" FIELD4="FIELD4_VALUE" FIELD5="FIELD5_VALUE" FIELD6="FIELD6_VALUE" FIELD7="FIELD7_VALUE" FIELD8="FIELD8_VALUE" FIELD9="FIELD9_VALUE" FIELD10="FIELD10_VALUE" FIELD11="FIELD11_VALUE" FIELD12="FIELD12_VALUE" FIELD13="FIELD13_VALUE" FIELD14="FIELD14_VALUE" FIELD15="FIELD15_VALUE" FIELD16="FIELD16_VALUE" FIELD17="FIELD17_VALUE" FIELD18="FIELD18_VALUE" FIELD19="FIELD19_VALUE" FIELD20="FIELD20_VALUE">
+      <LEVEL1 parent="//@rootSegment" document="/" CHAR_FIELD="1234ABCDEF" QUAN_FIELD="1234567890123456789" UNIT_FIELD="LBS" NUMC_FIELD="1234567890" DATS_FIELD="1863-07-03T06:00:00.000+0100" TIMS_FIELD="1970-01-01T18:15:30.000+0100" CURR_FIELD="1234567890123456789" CUKY_FIELD="USD" LANG_FIELD="EN" CLNT_FIELD="100" INT1_FIELD="255" INT2_FIELD="65535" INT4_FIELD="4294967295" FLTP_FIELD="250000000000000" ACCP_FIELD="186307" PREC_FIELD="12" LRAW_FIELD="0F0E0D0C0B0A09080706050403020100" DEC_FIELD="1234567890" RAW_FIELD="0F0E0D0C0B0A09080706050403020100" STRING_FIELD="01234567890ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz" RAWSTRING_FIELD="0F0E0D0C0B0A09080706050403020100">
         <segmentChildren parent="//@rootSegment/@segmentChildren/@LEVEL1.0">
-          <LEVEL2 parent="//@rootSegment/@segmentChildren/@LEVEL1.0" document="/" FIELD0="FIELD0_VALUE" FIELD1="FIELD1_VALUE" FIELD2="FIELD2_VALUE" FIELD3="FIELD3_VALUE" FIELD4="FIELD4_VALUE" FIELD5="FIELD5_VALUE" FIELD6="FIELD6_VALUE" FIELD7="FIELD7_VALUE" FIELD8="FIELD8_VALUE" FIELD9="FIELD9_VALUE" FIELD10="FIELD10_VALUE" FIELD11="FIELD11_VALUE" FIELD12="FIELD12_VALUE" FIELD13="FIELD13_VALUE" FIELD14="FIELD14_VALUE" FIELD15="FIELD15_VALUE" FIELD16="FIELD16_VALUE" FIELD17="FIELD17_VALUE" FIELD18="FIELD18_VALUE" FIELD19="FIELD19_VALUE" FIELD20="FIELD20_VALUE">
+          <LEVEL2 parent="//@rootSegment/@segmentChildren/@LEVEL1.0" document="/" CHAR_FIELD="1234ABCDEF" QUAN_FIELD="1234567890123456789" UNIT_FIELD="LBS" NUMC_FIELD="1234567890" DATS_FIELD="1863-07-03T06:00:00.000+0100" TIMS_FIELD="1970-01-01T18:15:30.000+0100" CURR_FIELD="1234567890123456789" CUKY_FIELD="USD" LANG_FIELD="EN" CLNT_FIELD="100" INT1_FIELD="255" INT2_FIELD="65535" INT4_FIELD="4294967295" FLTP_FIELD="250000000000000" ACCP_FIELD="186307" PREC_FIELD="12" LRAW_FIELD="0F0E0D0C0B0A09080706050403020100" DEC_FIELD="1234567890" RAW_FIELD="0F0E0D0C0B0A09080706050403020100" STRING_FIELD="01234567890ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz" RAWSTRING_FIELD="0F0E0D0C0B0A09080706050403020100">
             <segmentChildren parent="//@rootSegment/@segmentChildren/@LEVEL1.0/@segmentChildren/@LEVEL2.0">
-              <LEVEL3 parent="//@rootSegment/@segmentChildren/@LEVEL1.0/@segmentChildren/@LEVEL2.0" document="/" FIELD0="FIELD0_VALUE" FIELD1="FIELD1_VALUE" FIELD2="FIELD2_VALUE" FIELD3="FIELD3_VALUE" FIELD4="FIELD4_VALUE" FIELD5="FIELD5_VALUE" FIELD6="FIELD6_VALUE" FIELD7="FIELD7_VALUE" FIELD8="FIELD8_VALUE" FIELD9="FIELD9_VALUE" FIELD10="FIELD10_VALUE" FIELD11="FIELD11_VALUE" FIELD12="FIELD12_VALUE" FIELD13="FIELD13_VALUE" FIELD14="FIELD14_VALUE" FIELD15="FIELD15_VALUE" FIELD16="FIELD16_VALUE" FIELD17="FIELD17_VALUE" FIELD18="FIELD18_VALUE" FIELD19="FIELD19_VALUE" FIELD20="FIELD20_VALUE"/>
+              <LEVEL3 parent="//@rootSegment/@segmentChildren/@LEVEL1.0/@segmentChildren/@LEVEL2.0" document="/" CHAR_FIELD="1234ABCDEF" QUAN_FIELD="1234567890123456789" UNIT_FIELD="LBS" NUMC_FIELD="1234567890" DATS_FIELD="1863-07-03T06:00:00.000+0100" TIMS_FIELD="1970-01-01T18:15:30.000+0100" CURR_FIELD="1234567890123456789" CUKY_FIELD="USD" LANG_FIELD="EN" CLNT_FIELD="100" INT1_FIELD="255" INT2_FIELD="65535" INT4_FIELD="4294967295" FLTP_FIELD="250000000000000" ACCP_FIELD="186307" PREC_FIELD="12" LRAW_FIELD="0F0E0D0C0B0A09080706050403020100" DEC_FIELD="1234567890" RAW_FIELD="0F0E0D0C0B0A09080706050403020100" STRING_FIELD="01234567890ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz" RAWSTRING_FIELD="0F0E0D0C0B0A09080706050403020100"/>
             </segmentChildren>
           </LEVEL2>
         </segmentChildren>
