+++ camel-sap/camel-sap-component/data/testRequest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="ASCII"?>
-<TEST_FUNCTION_MODULE:Request xmlns:TEST_FUNCTION_MODULE="http://sap.fusesource.org/rfc/TEST_REPOSITORY/TEST_FUNCTION_MODULE" PARAM_LIST_CHAR_PARAM="ABCDEFGHIJ" PARAM_LIST_NUM_PARAM="0123456789" PARAM_LIST_INT_PARAM="1968526677" PARAM_LIST_FLOAT_PARAM="1.0E38" PARAM_LIST_BCD_PARAM="100.00000000000001" PARAM_LIST_BINARY_PARAM="55" PARAM_LIST_BINARY_ARRAY_PARAM="FF0F1E2D3C4B5A607988" PARAM_LIST_DATE_PARAM="1861-04-12T00:00:00.000-0500" PARAM_LIST_TIME_PARAM="1970-01-01T12:15:30.000-0500" PARAM_LIST_STRING_PARAM="Four score and seven years ago ...">
-  <PARAM_LIST_STRUCTURE_PARAM CHAR_PARAM="ABCDEFGHIJ" NUM_PARAM="0123456789" INT_PARAM="1968526677" FLOAT_PARAM="1.0E38" BCD_PARAM="100.00000000000001" BINARY_PARAM="55" BINARY_ARRAY_PARAM="FF0F1E2D3C4B5A607988" DATE_PARAM="1861-04-12T00:00:00.000-0500" TIME_PARAM="1970-01-01T12:15:30.000-0500" STRING_PARAM="Four score and seven years ago ..."/>
+<TEST_FUNCTION_MODULE:Request xmlns:TEST_FUNCTION_MODULE="http://sap.fusesource.org/rfc/TEST_REPOSITORY/TEST_FUNCTION_MODULE" PARAM_LIST_CHAR_PARAM="ABCDEFGHIJ" PARAM_LIST_NUM_PARAM="0123456789" PARAM_LIST_INT_PARAM="1968526677" PARAM_LIST_FLOAT_PARAM="1.0E38" PARAM_LIST_BCD_PARAM="100.00000000000001" PARAM_LIST_BINARY_PARAM="55" PARAM_LIST_BINARY_ARRAY_PARAM="FF0F1E2D3C4B5A607988" PARAM_LIST_DATE_PARAM="1861-04-12T00:00:00.000+0100" PARAM_LIST_TIME_PARAM="1970-01-01T12:15:30.000+0100" PARAM_LIST_STRING_PARAM="Four score and seven years ago ...">
+  <PARAM_LIST_STRUCTURE_PARAM CHAR_PARAM="ABCDEFGHIJ" NUM_PARAM="0123456789" INT_PARAM="1968526677" FLOAT_PARAM="1.0E38" BCD_PARAM="100.00000000000001" BINARY_PARAM="55" BINARY_ARRAY_PARAM="FF0F1E2D3C4B5A607988" DATE_PARAM="1861-04-12T00:00:00.000+0100" TIME_PARAM="1970-01-01T12:15:30.000+0100" STRING_PARAM="Four score and seven years ago ..."/>
   <PARAM_LIST_TABLE_PARAM>
-    <row CHAR_PARAM="ABCDEFGHIJ" NUM_PARAM="0123456789" INT_PARAM="1968526677" FLOAT_PARAM="1.0E38" BCD_PARAM="100.00000000000001" BINARY_PARAM="55" BINARY_ARRAY_PARAM="FF0F1E2D3C4B5A607988" DATE_PARAM="1861-04-12T00:00:00.000-0500" TIME_PARAM="1970-01-01T12:15:30.000-0500" STRING_PARAM="Four score and seven years ago ..."/>
+    <row CHAR_PARAM="ABCDEFGHIJ" NUM_PARAM="0123456789" INT_PARAM="1968526677" FLOAT_PARAM="1.0E38" BCD_PARAM="100.00000000000001" BINARY_PARAM="55" BINARY_ARRAY_PARAM="FF0F1E2D3C4B5A607988" DATE_PARAM="1861-04-12T00:00:00.000+0100" TIME_PARAM="1970-01-01T12:15:30.000+0100" STRING_PARAM="Four score and seven years ago ..."/>
   </PARAM_LIST_TABLE_PARAM>
 </TEST_FUNCTION_MODULE:Request>
