+++ camel-sap/camel-sap-component/src/test/java/org/fusesource/camel/component/sap/SapQueuedRfcProducerTest.java
@@ -17,30 +17,21 @@
 package org.fusesource.camel.component.sap;
 
 
+import java.util.Map;
+import java.util.Properties;
+import com.sap.conn.jco.JCoDestinationManager;
 import org.apache.camel.Exchange;
 import org.apache.camel.builder.RouteBuilder;
 import org.fusesource.camel.component.sap.model.rfc.Structure;
 import org.junit.jupiter.api.Test;
-
+import org.mockito.MockedStatic;
 import org.mockito.Mockito;
-import org.powermock.api.mockito.PowerMockito;
-import org.powermock.api.mockito.mockpolicies.Slf4jMockPolicy;
-import org.powermock.core.classloader.annotations.MockPolicy;
-import org.powermock.core.classloader.annotations.PrepareForTest;
-import org.powermock.modules.junit4.PowerMockRunner;
 
-import com.sap.conn.jco.JCoDestinationManager;
-
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import org.mockito.MockedStatic;
-
-import java.util.Map;
-import java.util.Properties;
 
 /**
  * SAP Producer test cases.
